Combine UserInfo KcOidcBrokerConfiguration with OidcUserInfoClaimToRoleMapperTest
